server/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
boto3==1.34.4
botocore==1.34.5
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
django-redis==5.4.0
djongo==1.3.6
dnspython==2.4.2
idna==3.6
jmespath==1.0.1
lxml==4.9.4
mysqlclient==2.2.1
numpy==1.26.2
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
drf-yasg==1.21.7